requirements.txt
groq
python-dotenv
gradio
torchaudio
transformers
yt-dlp
ffmpeg-python